+++ packages/dnd/src/QueryBuilderDnD.test.tsx
@@ -1,8 +1,9 @@
+import { consoleMocks, userEventSetup } from '@rqb-testing';
 import { act, render, screen } from '@testing-library/react';
 import * as React from 'react';
+import * as reactDnD from 'react-dnd';
 import * as reactDnDHTML5Backend from 'react-dnd-html5-backend';
 import { simulateDragDrop, simulateDragHover, wrapWithTestBackend } from 'react-dnd-test-utils';
-import * as reactDnD from 'react-dnd';
 import type {
   Field,
   FullCombinator,
@@ -18,8 +19,9 @@ import {
   getCompatContextProvider,
   standardClassnames,
 } from 'react-querybuilder';
-import { consoleMocks } from '@rqb-testing';
 import { QueryBuilderDnD, QueryBuilderDndWithoutProvider } from './QueryBuilderDnD';
+
+const user = userEventSetup();
 
 const getHandlerId = (el: HTMLElement, dragDrop: 'drag' | 'drop') => () =>
   el.getAttribute(`data-${dragDrop}monitorid`);
@@ -142,11 +144,13 @@ describe.each([{ QBctx: QueryBuilderDnD }, { QBctx: QueryBuilderDndWithoutProvid
         const rules = screen.getAllByTestId(TestID.rule);
         simulateDragDrop(getHandlerId(rules[0], 'drag'), getHandlerId(rules[1], 'drop'), gDnDBe());
         expect(onQueryChange).toHaveBeenCalledTimes(2);
-        expect(onQueryChange).toHaveBeenLastCalledWith(
-          expect.objectContaining({
-            rules: [expect.objectContaining({ id: '1' }), expect.objectContaining({ id: '0' })],
-          })
-        );
+        expect(onQueryChange).toHaveBeenLastCalledWith({
+          combinator: 'and',
+          rules: [
+            { id: '1', field: 'field1', operator: '=', value: '1' },
+            { id: '0', field: 'field0', operator: '=', value: '0' },
+          ],
+        });
       });
 
       it('moves a rule to a different group with a common ancestor', () => {
@@ -174,20 +178,23 @@ describe.each([{ QBctx: QueryBuilderDnD }, { QBctx: QueryBuilderDndWithoutProvid
         const ruleGroup = screen.getAllByTestId(TestID.ruleGroup)[2]; // id 3
         simulateDragDrop(getHandlerId(rule, 'drag'), getHandlerId(ruleGroup, 'drop'), gDnDBe());
         expect(onQueryChange).toHaveBeenCalledTimes(2);
-        expect(onQueryChange).toHaveBeenLastCalledWith(
-          expect.objectContaining({
-            rules: [
-              expect.objectContaining({
-                rules: expect.arrayContaining([
-                  expect.anything(),
-                  expect.objectContaining({
-                    rules: [expect.objectContaining({ id: '2' })],
-                  }),
-                ]),
-              }),
-            ],
-          })
-        );
+        expect(onQueryChange).toHaveBeenLastCalledWith({
+          combinator: 'and',
+          rules: [
+            {
+              id: '0',
+              combinator: 'and',
+              rules: [
+                { id: '1', field: 'field0', operator: '=', value: '1' },
+                {
+                  id: '3',
+                  combinator: 'and',
+                  rules: [{ id: '2', field: 'field0', operator: '=', value: '2' }],
+                },
+              ],
+            },
+          ],
+        });
       });
     });
 
@@ -213,15 +220,13 @@ describe.each([{ QBctx: QueryBuilderDnD }, { QBctx: QueryBuilderDndWithoutProvid
           gDnDBeIC()
         );
         expect(onQueryChange).toHaveBeenCalledTimes(2);
-        expect(onQueryChange).toHaveBeenLastCalledWith(
-          expect.objectContaining({
-            rules: [
-              expect.objectContaining({ field: 'field1', operator: '=', value: '1' }),
-              'and',
-              expect.objectContaining({ field: 'field0', operator: '=', value: '0' }),
-            ],
-          })
-        );
+        expect(onQueryChange).toHaveBeenLastCalledWith({
+          rules: [
+            { field: 'field1', operator: '=', value: '1' },
+            'and',
+            { field: 'field0', operator: '=', value: '0' },
+          ],
+        });
       });
 
       it('swaps the last rule with the first within the same group', () => {
@@ -246,15 +251,13 @@ describe.each([{ QBctx: QueryBuilderDnD }, { QBctx: QueryBuilderDndWithoutProvid
           gDnDBeIC()
         );
         expect(onQueryChange).toHaveBeenCalledTimes(2);
-        expect(onQueryChange).toHaveBeenLastCalledWith(
-          expect.objectContaining({
-            rules: [
-              expect.objectContaining({ field: 'field1', operator: '=', value: '1' }),
-              'and',
-              expect.objectContaining({ field: 'field0', operator: '=', value: '0' }),
-            ],
-          })
-        );
+        expect(onQueryChange).toHaveBeenLastCalledWith({
+          rules: [
+            { field: 'field1', operator: '=', value: '1' },
+            'and',
+            { field: 'field0', operator: '=', value: '0' },
+          ],
+        });
       });
 
       it('moves a rule from first to last within the same group', () => {
@@ -280,17 +283,15 @@ describe.each([{ QBctx: QueryBuilderDnD }, { QBctx: QueryBuilderDndWithoutProvid
           gDnDBeIC()
         );
         expect(onQueryChange).toHaveBeenCalledTimes(2);
-        expect(onQueryChange).toHaveBeenLastCalledWith(
-          expect.objectContaining({
-            rules: [
-              expect.objectContaining({ field: 'field1', operator: '=', value: '1' }),
-              'and',
-              expect.objectContaining({ field: 'field2', operator: '=', value: '2' }),
-              'and',
-              expect.objectContaining({ field: 'field0', operator: '=', value: '0' }),
-            ],
-          })
-        );
+        expect(onQueryChange).toHaveBeenLastCalledWith({
+          rules: [
+            { field: 'field1', operator: '=', value: '1' },
+            'and',
+            { field: 'field2', operator: '=', value: '2' },
+            'and',
+            { field: 'field0', operator: '=', value: '0' },
+          ],
+        });
       });
 
       it('moves a rule from last to first within the same group', () => {
@@ -317,17 +318,15 @@ describe.each([{ QBctx: QueryBuilderDnD }, { QBctx: QueryBuilderDndWithoutProvid
           gDnDBeIC()
         );
         expect(onQueryChange).toHaveBeenCalledTimes(2);
-        expect(onQueryChange).toHaveBeenLastCalledWith(
-          expect.objectContaining({
-            rules: [
-              expect.objectContaining({ field: 'field2', operator: '=', value: '2' }),
-              'and',
-              expect.objectContaining({ field: 'field0', operator: '=', value: '0' }),
-              'and',
-              expect.objectContaining({ field: 'field1', operator: '=', value: '1' }),
-            ],
-          })
-        );
+        expect(onQueryChange).toHaveBeenLastCalledWith({
+          rules: [
+            { field: 'field2', operator: '=', value: '2' },
+            'and',
+            { field: 'field0', operator: '=', value: '0' },
+            'and',
+            { field: 'field1', operator: '=', value: '1' },
+          ],
+        });
       });
 
       it('moves a rule from last to middle by dropping on inline combinator', () => {
@@ -354,17 +353,15 @@ describe.each([{ QBctx: QueryBuilderDnD }, { QBctx: QueryBuilderDndWithoutProvid
           gDnDBeIC()
         );
         expect(onQueryChange).toHaveBeenCalledTimes(2);
-        expect(onQueryChange).toHaveBeenLastCalledWith(
-          expect.objectContaining({
-            rules: [
-              expect.objectContaining({ field: 'field0', operator: '=', value: '0' }),
-              'and',
-              expect.objectContaining({ field: 'field2', operator: '=', value: '2' }),
-              'and',
-              expect.objectContaining({ field: 'field1', operator: '=', value: '1' }),
-            ],
-          })
-        );
+        expect(onQueryChange).toHaveBeenLastCalledWith({
+          rules: [
+            { field: 'field0', operator: '=', value: '0' },
+            'and',
+            { field: 'field2', operator: '=', value: '2' },
+            'and',
+            { field: 'field1', operator: '=', value: '1' },
+          ],
+        });
       });
 
       it('moves a first-child rule to a different group as the first child', () => {
@@ -391,21 +388,19 @@ describe.each([{ QBctx: QueryBuilderDnD }, { QBctx: QueryBuilderDndWithoutProvid
         const ruleGroup = screen.getAllByTestId(TestID.ruleGroup)[1];
         simulateDragDrop(getHandlerId(rule, 'drag'), getHandlerId(ruleGroup, 'drop'), gDnDBeIC());
         expect(onQueryChange).toHaveBeenCalledTimes(2);
-        expect(onQueryChange).toHaveBeenLastCalledWith(
-          expect.objectContaining({
-            rules: [
-              expect.objectContaining({
-                rules: [
-                  expect.objectContaining({ field: 'field0', operator: '=', value: '0' }),
-                  'and',
-                  expect.objectContaining({ field: 'field1', operator: '=', value: '1' }),
-                  'and',
-                  expect.objectContaining({ field: 'field2', operator: '=', value: '2' }),
-                ],
-              }),
-            ],
-          })
-        );
+        expect(onQueryChange).toHaveBeenLastCalledWith({
+          rules: [
+            {
+              rules: [
+                { field: 'field0', operator: '=', value: '0' },
+                'and',
+                { field: 'field1', operator: '=', value: '1' },
+                'and',
+                { field: 'field2', operator: '=', value: '2' },
+              ],
+            },
+          ],
+        });
       });
 
       it('moves a middle-child rule to a different group as a middle child', () => {
@@ -439,25 +434,23 @@ describe.each([{ QBctx: QueryBuilderDnD }, { QBctx: QueryBuilderDndWithoutProvid
           gDnDBeIC()
         );
         expect(onQueryChange).toHaveBeenCalledTimes(2);
-        expect(onQueryChange).toHaveBeenLastCalledWith(
-          expect.objectContaining({
-            rules: [
-              expect.objectContaining({ field: 'field0', operator: '=', value: '0' }),
-              'and',
-              expect.objectContaining({ field: 'field2', operator: '=', value: '2' }),
-              'and',
-              expect.objectContaining({
-                rules: [
-                  expect.objectContaining({ field: 'field3', operator: '=', value: '3' }),
-                  'and',
-                  expect.objectContaining({ field: 'field1', operator: '=', value: '1' }),
-                  'and',
-                  expect.objectContaining({ field: 'field4', operator: '=', value: '4' }),
-                ],
-              }),
-            ],
-          })
-        );
+        expect(onQueryChange).toHaveBeenLastCalledWith({
+          rules: [
+            { field: 'field0', operator: '=', value: '0' },
+            'and',
+            { field: 'field2', operator: '=', value: '2' },
+            'and',
+            {
+              rules: [
+                { field: 'field3', operator: '=', value: '3' },
+                'and',
+                { field: 'field1', operator: '=', value: '1' },
+                'and',
+                { field: 'field4', operator: '=', value: '4' },
+              ],
+            },
+          ],
+        });
       });
     });
   }
@@ -650,22 +643,94 @@ it('can move rules/groups to different query builders', async () => {
     getHandlerId(dropRule, 'drop'),
     getDndBackend()!
   );
-  expect(onQueryChange).toHaveBeenCalledWith(
-    expect.objectContaining({
-      combinator: 'and',
-      rules: expect.arrayContaining([
-        expect.objectContaining({ field: 'field1', operator: '=', value: '1' }),
-        expect.objectContaining({ field: 'field4', operator: '=', value: '4' }),
-        expect.objectContaining({ field: 'field2', operator: '=', value: '2' }),
-      ]),
-    })
+  expect(onQueryChange).toHaveBeenCalledWith({
+    combinator: 'and',
+    rules: [
+      { field: 'field1', operator: '=', value: '1' },
+      {
+        field: 'field4',
+        operator: '=',
+        value: '4',
+        id: expect.any(String),
+        path: expect.arrayContaining([expect.any(Number)]),
+        qbId: expect.any(String),
+      },
+      { field: 'field2', operator: '=', value: '2' },
+    ],
+  });
+  expect(onQueryChange).toHaveBeenCalledWith({
+    combinator: 'and',
+    rules: [{ field: 'field3', operator: '=', value: '3' }],
+  });
+});
+
+it('can group rules/groups to different query builders', async () => {
+  const onQueryChange = jest.fn<never, [RuleGroupType]>();
+  const fields: Field[] = [
+    { name: 'field1', label: 'Field 1' },
+    { name: 'field2', label: 'Field 2' },
+    { name: 'field3', label: 'Field 3' },
+    { name: 'field4', label: 'Field 4' },
+  ];
+  const query1: RuleGroupType = {
+    combinator: 'and',
+    rules: [
+      { field: 'field1', operator: '=', value: '1' },
+      { field: 'field2', operator: '=', value: '2' },
+    ],
+  };
+  const query2: RuleGroupType = {
+    combinator: 'and',
+    rules: [
+      { field: 'field3', operator: '=', value: '3' },
+      { field: 'field4', operator: '=', value: '4' },
+    ],
+  };
+  const [QueryBuilderWrapped, getDndBackend] = wrapWithTestBackend(
+    (props: QueryBuilderProps<RuleGroupType, FullField, FullOperator, FullCombinator>) => (
+      <QueryBuilderDnD dnd={{ ...reactDnD, ...reactDnDHTML5Backend }}>
+        <QueryBuilder {...props} query={query1} />
+        <QueryBuilder {...props} query={query2} />
+      </QueryBuilderDnD>
+    )
   );
-  expect(onQueryChange).toHaveBeenCalledWith(
-    expect.objectContaining({
-      combinator: 'and',
-      rules: expect.arrayContaining([
-        expect.objectContaining({ field: 'field1', operator: '=', value: '1' }),
-      ]),
-    })
+  render(<QueryBuilderWrapped fields={fields} enableDragAndDrop onQueryChange={onQueryChange} />);
+  const [dropRule, _1, _2, dragRule] = screen.getAllByTestId(TestID.rule);
+  await user.keyboard('{Control>}');
+  simulateDragDrop(
+    getHandlerId(dragRule, 'drag'),
+    getHandlerId(dropRule, 'drop'),
+    getDndBackend()!
   );
+  await user.keyboard('{/Control}');
+  expect(onQueryChange).toHaveBeenCalledWith({
+    combinator: 'and',
+    rules: [
+      {
+        id: expect.any(String),
+        combinator: 'and',
+        rules: [
+          {
+            id: expect.any(String),
+            field: 'field1',
+            operator: '=',
+            value: '1',
+          },
+          {
+            field: 'field4',
+            operator: '=',
+            value: '4',
+            id: expect.any(String),
+            path: expect.arrayContaining([expect.any(Number)]),
+            qbId: expect.any(String),
+          },
+        ],
+      },
+      { field: 'field2', operator: '=', value: '2' },
+    ],
+  });
+  expect(onQueryChange).toHaveBeenCalledWith({
+    combinator: 'and',
+    rules: [{ field: 'field3', operator: '=', value: '3' }],
+  });
 });